combine model-deployer with served-models for CLI
